+++ inst/doc/introduction.R
@@ -1,25 +1,28 @@
-## ----install github, warning = FALSE, results = "hide", message = FALSE, eval = FALSE----
-#  devtools::install_github("jandix/rzeit2")
+## ---- eval=FALSE---------------------------------------------------------
+#  # save the api key in the .Renviron file
+#  set_api_key(api_key = "xxx",
+#              path = "~/.Renviron")
 
-## ----loading rzeit, warning = FALSE, results = "hide", message = FALSE, eval = FALSE----
-#  library(rzeit2)
-
-## ----set key, warning = FALSE, message = FALSE, eval = FALSE-------------
-#  set_api_key(api_key = "set_your_api_key_here",
-#              path = tempdir())
-
-## ----get content, warning = FALSE, message = FALSE, eval = FALSE---------
-#  articles_merkel <- get_content("Merkel",
-#                                 limit = 100,
+## ---- eval=FALSE---------------------------------------------------------
+#  # fetch articles up to 1000 rows
+#  tatort_articles <- get_content(query = "Tatort",
 #                                 begin_date = "20180101",
-#                                 end_date = "20180520")
+#                                 end_date = "20180131")
+#  
+#  # fetch ALL articles
+#  tatort_articles <- get_content(query = "Tatort",
+#                                 timeout = 2)
 
-## ----get content all, warning = FALSE, message = FALSE, eval = FALSE-----
-#  articles_merkel <- get_content_all("Merkel",
-#                                     timeout = 1,
-#                                     begin_date = "20150101",
-#                                     end_date = "20180520")
+## ---- eval=FALSE---------------------------------------------------------
+#  tatort_content <- get_article_text(url = tatort_articles$content$href,
+#                                     timeout = 1)
 
-## ----get client, warning = FALSE, message = FALSE, eval = FALSE----------
-#  get_client()
+## ---- eval=FALSE---------------------------------------------------------
+#  tatort_comments <- get_article_comments(url = tatort_articles$content$href[1],
+#                                          timeout = 1)
+
+## ---- eval=FALSE---------------------------------------------------------
+#  tatort_images <- get_article_images(url = tatort_articles$content$href,
+#                                      timeout = 1,
+#                                      download = "~/Documents/tatort-img/")