+++ NEWS.md
@@ -1,3 +1,10 @@
+# version 0.2.0 (2018-10-18)
+
+- define new helper functions
+- new function to scrape images from articles (get_article_images)
+- new function to scrape comments from articles (get_article_comments)
+- update introduction vignette
+
 # version 0.1.0 (2018-06-26)
 
 - initial publication on Github